renamed bookings to partyBookings
