+++ firebase/functions/fizz-kidz/src/onboarding/index.ts
@@ -1,4 +1,4 @@
-import { Locations } from '../booking/Locations'
+import { Locations } from '../partyBookings/Locations'
 
 export type Employee = {
     id: string
