+++ firebase/functions/fizz-kidz/src/partyBookings/Event.ts
