+++ firebase/functions/src/partyBookings/core/utils.ts
