+++ firebase/functions/src/partyBookings/functions/deletePartyBooking.ts
