+++ firebase/functions/src/partyBookings/functions/updatePartyBooking.ts
